+++ src/__generated__/ArtistArtworkGrid_artwork.graphql.ts
@@ -1,0 +1,185 @@
+/* tslint:disable */
+
+import { ReaderFragment } from "relay-runtime";
+import { GenericGrid_artworks$ref } from "./GenericGrid_artworks.graphql";
+declare const _ArtistArtworkGrid_artwork$ref: unique symbol;
+export type ArtistArtworkGrid_artwork$ref = typeof _ArtistArtworkGrid_artwork$ref;
+export type ArtistArtworkGrid_artwork = {
+    readonly id: string;
+    readonly artist: {
+        readonly name: string | null;
+        readonly artworks_connection: {
+            readonly pageInfo: {
+                readonly startCursor: string | null;
+                readonly endCursor: string | null;
+                readonly hasPreviousPage: boolean;
+                readonly hasNextPage: boolean;
+            };
+            readonly edges: ReadonlyArray<{
+                readonly node: {
+                    readonly " $fragmentRefs": GenericGrid_artworks$ref;
+                } | null;
+            } | null> | null;
+        } | null;
+    } | null;
+    readonly " $refType": ArtistArtworkGrid_artwork$ref;
+};
+
+
+
+const node: ReaderFragment = {
+  "kind": "Fragment",
+  "name": "ArtistArtworkGrid_artwork",
+  "type": "Artwork",
+  "metadata": {
+    "connection": [
+      {
+        "count": null,
+        "cursor": null,
+        "direction": "forward",
+        "path": [
+          "artist",
+          "artworks_connection"
+        ]
+      }
+    ]
+  },
+  "argumentDefinitions": [
+    {
+      "kind": "RootArgument",
+      "name": "excludeArtworkIds",
+      "type": "[String]"
+    }
+  ],
+  "selections": [
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "id",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "LinkedField",
+      "alias": null,
+      "name": "artist",
+      "storageKey": null,
+      "args": null,
+      "concreteType": "Artist",
+      "plural": false,
+      "selections": [
+        {
+          "kind": "ScalarField",
+          "alias": null,
+          "name": "name",
+          "args": null,
+          "storageKey": null
+        },
+        {
+          "kind": "LinkedField",
+          "alias": "artworks_connection",
+          "name": "__GenericGrid_artworks_connection_connection",
+          "storageKey": null,
+          "args": [
+            {
+              "kind": "Variable",
+              "name": "exclude",
+              "variableName": "excludeArtworkIds"
+            },
+            {
+              "kind": "Literal",
+              "name": "sort",
+              "value": "PUBLISHED_AT_DESC"
+            }
+          ],
+          "concreteType": "ArtworkConnection",
+          "plural": false,
+          "selections": [
+            {
+              "kind": "LinkedField",
+              "alias": null,
+              "name": "pageInfo",
+              "storageKey": null,
+              "args": null,
+              "concreteType": "PageInfo",
+              "plural": false,
+              "selections": [
+                {
+                  "kind": "ScalarField",
+                  "alias": null,
+                  "name": "startCursor",
+                  "args": null,
+                  "storageKey": null
+                },
+                {
+                  "kind": "ScalarField",
+                  "alias": null,
+                  "name": "endCursor",
+                  "args": null,
+                  "storageKey": null
+                },
+                {
+                  "kind": "ScalarField",
+                  "alias": null,
+                  "name": "hasPreviousPage",
+                  "args": null,
+                  "storageKey": null
+                },
+                {
+                  "kind": "ScalarField",
+                  "alias": null,
+                  "name": "hasNextPage",
+                  "args": null,
+                  "storageKey": null
+                }
+              ]
+            },
+            {
+              "kind": "LinkedField",
+              "alias": null,
+              "name": "edges",
+              "storageKey": null,
+              "args": null,
+              "concreteType": "ArtworkEdge",
+              "plural": true,
+              "selections": [
+                {
+                  "kind": "LinkedField",
+                  "alias": null,
+                  "name": "node",
+                  "storageKey": null,
+                  "args": null,
+                  "concreteType": "Artwork",
+                  "plural": false,
+                  "selections": [
+                    {
+                      "kind": "ScalarField",
+                      "alias": null,
+                      "name": "__typename",
+                      "args": null,
+                      "storageKey": null
+                    },
+                    {
+                      "kind": "FragmentSpread",
+                      "name": "GenericGrid_artworks",
+                      "args": null
+                    }
+                  ]
+                },
+                {
+                  "kind": "ScalarField",
+                  "alias": null,
+                  "name": "cursor",
+                  "args": null,
+                  "storageKey": null
+                }
+              ]
+            }
+          ]
+        }
+      ]
+    }
+  ]
+};
+(node as any).hash = '284c1be568c42a7cff2f2b79e651914d';
+export default node;