+++ src/__generated__/ArtworkContextArtistTestsQuery.graphql.ts
@@ -1,92 +1,30 @@
 /* tslint:disable */
 
 import { ConcreteRequest } from "relay-runtime";
-import { Artwork_artwork$ref } from "./Artwork_artwork.graphql";
-export type ArtworkQueryVariables = {
-    readonly artworkID: string;
-    readonly excludeArtworkIds?: ReadonlyArray<string> | null;
+import { ArtworkContextArtist_artwork$ref } from "./ArtworkContextArtist_artwork.graphql";
+export type ArtworkContextArtistTestsQueryVariables = {
+    readonly excludeArtworkIds?: ReadonlyArray<string | null> | null;
 };
-export type ArtworkQueryResponse = {
+export type ArtworkContextArtistTestsQueryResponse = {
     readonly artwork: {
-        readonly " $fragmentRefs": Artwork_artwork$ref;
+        readonly " $fragmentRefs": ArtworkContextArtist_artwork$ref;
     } | null;
 };
-export type ArtworkQuery = {
-    readonly response: ArtworkQueryResponse;
-    readonly variables: ArtworkQueryVariables;
+export type ArtworkContextArtistTestsQuery = {
+    readonly response: ArtworkContextArtistTestsQueryResponse;
+    readonly variables: ArtworkContextArtistTestsQueryVariables;
 };
 
 
 
 /*
-query ArtworkQuery(
-  $artworkID: String!
-  $excludeArtworkIds: [String!]
+query ArtworkContextArtistTestsQuery(
+  $excludeArtworkIds: [String]
 ) {
-  artwork(id: $artworkID) {
-    ...Artwork_artwork
+  artwork(id: "abbas-kiarostami-untitled-7") {
+    ...ArtworkContextArtist_artwork
     id
   }
-}
-
-fragment Artwork_artwork on Artwork {
-  ...ArtworkTombstone_artwork
-  ...ArtworkActions_artwork
-  ...ArtworkAvailability_artwork
-  ...SellerInfo_artwork
-  ...OtherWorks_artwork
-}
-
-fragment ArtworkTombstone_artwork on Artwork {
-  title
-  medium
-  date
-  cultural_maker
-  artists {
-    name
-    href
-    id
-  }
-  dimensions {
-    in
-    cm
-  }
-  edition_of
-  attribution_class {
-    short_description
-    id
-  }
-}
-
-fragment ArtworkActions_artwork on Artwork {
-  id
-  internalID
-  is_saved
-}
-
-fragment ArtworkAvailability_artwork on Artwork {
-  availability
-}
-
-fragment SellerInfo_artwork on Artwork {
-  partner {
-    name
-    id
-  }
-}
-
-fragment OtherWorks_artwork on Artwork {
-  id
-  context {
-    __typename
-    ... on Node {
-      id
-    }
-    ... on ArtworkContextFair {
-      id
-    }
-  }
-  ...ArtworkContextArtist_artwork
 }
 
 fragment ArtworkContextArtist_artwork on Artwork {
@@ -177,81 +115,33 @@ const node: ConcreteRequest = (function(){
 var v0 = [
   {
     "kind": "LocalArgument",
-    "name": "artworkID",
-    "type": "String!",
-    "defaultValue": null
-  },
-  {
-    "kind": "LocalArgument",
     "name": "excludeArtworkIds",
-    "type": "[String!]",
+    "type": "[String]",
     "defaultValue": null
   }
 ],
 v1 = [
   {
-    "kind": "Variable",
+    "kind": "Literal",
     "name": "id",
-    "variableName": "artworkID"
+    "value": "abbas-kiarostami-untitled-7"
   }
 ],
 v2 = {
-  "kind": "ScalarField",
-  "alias": null,
-  "name": "title",
-  "args": null,
-  "storageKey": null
-},
-v3 = {
-  "kind": "ScalarField",
-  "alias": null,
-  "name": "date",
-  "args": null,
-  "storageKey": null
-},
-v4 = {
-  "kind": "ScalarField",
-  "alias": null,
-  "name": "name",
-  "args": null,
-  "storageKey": null
-},
-v5 = {
-  "kind": "ScalarField",
-  "alias": null,
-  "name": "href",
-  "args": null,
-  "storageKey": null
-},
-v6 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "id",
   "args": null,
   "storageKey": null
 },
-v7 = [
-  (v4/*: any*/),
-  (v6/*: any*/)
-],
-v8 = {
-  "kind": "LinkedField",
-  "alias": null,
-  "name": "partner",
-  "storageKey": null,
-  "args": null,
-  "concreteType": "Partner",
-  "plural": false,
-  "selections": (v7/*: any*/)
-},
-v9 = {
+v3 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "__typename",
+  "name": "name",
   "args": null,
   "storageKey": null
 },
-v10 = [
+v4 = [
   {
     "kind": "Variable",
     "name": "exclude",
@@ -268,14 +158,14 @@ v10 = [
     "value": "PUBLISHED_AT_DESC"
   }
 ],
-v11 = {
+v5 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "is_closed",
   "args": null,
   "storageKey": null
 },
-v12 = [
+v6 = [
   {
     "kind": "ScalarField",
     "alias": null,
@@ -283,12 +173,16 @@ v12 = [
     "args": null,
     "storageKey": null
   }
+],
+v7 = [
+  (v3/*: any*/),
+  (v2/*: any*/)
 ];
 return {
   "kind": "Request",
   "fragment": {
     "kind": "Fragment",
-    "name": "ArtworkQuery",
+    "name": "ArtworkContextArtistTestsQuery",
     "type": "Query",
     "metadata": null,
     "argumentDefinitions": (v0/*: any*/),
@@ -297,14 +191,14 @@ return {
         "kind": "LinkedField",
         "alias": null,
         "name": "artwork",
-        "storageKey": null,
+        "storageKey": "artwork(id:\"abbas-kiarostami-untitled-7\")",
         "args": (v1/*: any*/),
         "concreteType": "Artwork",
         "plural": false,
         "selections": [
           {
             "kind": "FragmentSpread",
-            "name": "Artwork_artwork",
+            "name": "ArtworkContextArtist_artwork",
             "args": null
           }
         ]
@@ -313,135 +207,19 @@ return {
   },
   "operation": {
     "kind": "Operation",
-    "name": "ArtworkQuery",
+    "name": "ArtworkContextArtistTestsQuery",
     "argumentDefinitions": (v0/*: any*/),
     "selections": [
       {
         "kind": "LinkedField",
         "alias": null,
         "name": "artwork",
-        "storageKey": null,
+        "storageKey": "artwork(id:\"abbas-kiarostami-untitled-7\")",
         "args": (v1/*: any*/),
         "concreteType": "Artwork",
         "plural": false,
         "selections": [
           (v2/*: any*/),
-          {
-            "kind": "ScalarField",
-            "alias": null,
-            "name": "medium",
-            "args": null,
-            "storageKey": null
-          },
-          (v3/*: any*/),
-          {
-            "kind": "ScalarField",
-            "alias": null,
-            "name": "cultural_maker",
-            "args": null,
-            "storageKey": null
-          },
-          {
-            "kind": "LinkedField",
-            "alias": null,
-            "name": "artists",
-            "storageKey": null,
-            "args": null,
-            "concreteType": "Artist",
-            "plural": true,
-            "selections": [
-              (v4/*: any*/),
-              (v5/*: any*/),
-              (v6/*: any*/)
-            ]
-          },
-          {
-            "kind": "LinkedField",
-            "alias": null,
-            "name": "dimensions",
-            "storageKey": null,
-            "args": null,
-            "concreteType": "dimensions",
-            "plural": false,
-            "selections": [
-              {
-                "kind": "ScalarField",
-                "alias": null,
-                "name": "in",
-                "args": null,
-                "storageKey": null
-              },
-              {
-                "kind": "ScalarField",
-                "alias": null,
-                "name": "cm",
-                "args": null,
-                "storageKey": null
-              }
-            ]
-          },
-          {
-            "kind": "ScalarField",
-            "alias": null,
-            "name": "edition_of",
-            "args": null,
-            "storageKey": null
-          },
-          {
-            "kind": "LinkedField",
-            "alias": null,
-            "name": "attribution_class",
-            "storageKey": null,
-            "args": null,
-            "concreteType": "AttributionClass",
-            "plural": false,
-            "selections": [
-              {
-                "kind": "ScalarField",
-                "alias": null,
-                "name": "short_description",
-                "args": null,
-                "storageKey": null
-              },
-              (v6/*: any*/)
-            ]
-          },
-          (v6/*: any*/),
-          {
-            "kind": "ScalarField",
-            "alias": null,
-            "name": "internalID",
-            "args": null,
-            "storageKey": null
-          },
-          {
-            "kind": "ScalarField",
-            "alias": null,
-            "name": "is_saved",
-            "args": null,
-            "storageKey": null
-          },
-          {
-            "kind": "ScalarField",
-            "alias": null,
-            "name": "availability",
-            "args": null,
-            "storageKey": null
-          },
-          (v8/*: any*/),
-          {
-            "kind": "LinkedField",
-            "alias": null,
-            "name": "context",
-            "storageKey": null,
-            "args": null,
-            "concreteType": null,
-            "plural": false,
-            "selections": [
-              (v9/*: any*/),
-              (v6/*: any*/)
-            ]
-          },
           {
             "kind": "LinkedField",
             "alias": null,
@@ -451,13 +229,13 @@ return {
             "concreteType": "Artist",
             "plural": false,
             "selections": [
-              (v4/*: any*/),
+              (v3/*: any*/),
               {
                 "kind": "LinkedField",
                 "alias": null,
                 "name": "artworks_connection",
                 "storageKey": null,
-                "args": (v10/*: any*/),
+                "args": (v4/*: any*/),
                 "concreteType": "ArtworkConnection",
                 "plural": false,
                 "selections": [
@@ -518,7 +296,7 @@ return {
                         "concreteType": "Artwork",
                         "plural": false,
                         "selections": [
-                          (v6/*: any*/),
+                          (v2/*: any*/),
                           {
                             "kind": "ScalarField",
                             "alias": null,
@@ -557,8 +335,20 @@ return {
                               }
                             ]
                           },
-                          (v2/*: any*/),
-                          (v3/*: any*/),
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "title",
+                            "args": null,
+                            "storageKey": null
+                          },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "date",
+                            "args": null,
+                            "storageKey": null
+                          },
                           {
                             "kind": "ScalarField",
                             "alias": null,
@@ -624,7 +414,7 @@ return {
                                 "args": null,
                                 "storageKey": null
                               },
-                              (v11/*: any*/),
+                              (v5/*: any*/),
                               {
                                 "kind": "ScalarField",
                                 "alias": null,
@@ -632,7 +422,7 @@ return {
                                 "args": null,
                                 "storageKey": null
                               },
-                              (v6/*: any*/)
+                              (v2/*: any*/)
                             ]
                           },
                           {
@@ -652,7 +442,7 @@ return {
                                 "args": null,
                                 "concreteType": "SaleArtworkOpeningBid",
                                 "plural": false,
-                                "selections": (v12/*: any*/)
+                                "selections": (v6/*: any*/)
                               },
                               {
                                 "kind": "LinkedField",
@@ -662,7 +452,7 @@ return {
                                 "args": null,
                                 "concreteType": "SaleArtworkCurrentBid",
                                 "plural": false,
-                                "selections": (v12/*: any*/)
+                                "selections": (v6/*: any*/)
                               },
                               {
                                 "kind": "ScalarField",
@@ -680,11 +470,11 @@ return {
                                 "concreteType": "Sale",
                                 "plural": false,
                                 "selections": [
-                                  (v11/*: any*/),
-                                  (v6/*: any*/)
+                                  (v5/*: any*/),
+                                  (v2/*: any*/)
                                 ]
                               },
-                              (v6/*: any*/)
+                              (v2/*: any*/)
                             ]
                           },
                           {
@@ -703,9 +493,30 @@ return {
                             "plural": true,
                             "selections": (v7/*: any*/)
                           },
-                          (v8/*: any*/),
-                          (v5/*: any*/),
-                          (v9/*: any*/)
+                          {
+                            "kind": "LinkedField",
+                            "alias": null,
+                            "name": "partner",
+                            "storageKey": null,
+                            "args": null,
+                            "concreteType": "Partner",
+                            "plural": false,
+                            "selections": (v7/*: any*/)
+                          },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "href",
+                            "args": null,
+                            "storageKey": null
+                          },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "__typename",
+                            "args": null,
+                            "storageKey": null
+                          }
                         ]
                       },
                       {
@@ -723,7 +534,7 @@ return {
                 "kind": "LinkedHandle",
                 "alias": null,
                 "name": "artworks_connection",
-                "args": (v10/*: any*/),
+                "args": (v4/*: any*/),
                 "handle": "connection",
                 "key": "GenericGrid_artworks_connection",
                 "filters": [
@@ -731,7 +542,7 @@ return {
                   "exclude"
                 ]
               },
-              (v6/*: any*/)
+              (v2/*: any*/)
             ]
           }
         ]
@@ -740,12 +551,12 @@ return {
   },
   "params": {
     "operationKind": "query",
-    "name": "ArtworkQuery",
-    "id": "c111d05d25059025974d3b8ae652e1be",
+    "name": "ArtworkContextArtistTestsQuery",
+    "id": "0351d450b22b353ebf5535c01141c0c1",
     "text": null,
     "metadata": {}
   }
 };
 })();
-(node as any).hash = '8c594e52f5c996aecbaac4b08d7da014';
+(node as any).hash = '260f998e90def0188b619ca86b5ff2e5';
 export default node;