+++ src/lib/Scenes/Artwork/Components/OtherWorks/ArtworkContexts/ArtworkGrids/index.tsx
@@ -1,0 +1,1 @@
+export { ArtistArtworkGridFragmentContainer as ArtistArtworkGrid } from "./ArtistArtworkGrid"