+++ src/lib/__fixtures__/ArtworkFixture.ts
@@ -1,0 +1,93 @@
+export const ArtworkFixture = {
+  " $refType": null,
+  " $fragmentRefs": null,
+  id: "abbas-kiarostami-untitled-7",
+  _id: "5b2b745e9c18db204fc32e11",
+  __id: "QXJ0d29yazphYmJhcy1raWFyb3N0YW1pLXVudGl0bGVkLTc=",
+  title: "Untitled",
+  partner: {
+    id: "cama-gallery-1",
+    _id: "59ad77c6b202a34d6e059676",
+    name: "CAMA Gallery",
+  },
+  context: {
+    __typename: "ArtworkContextPartnerShow",
+  },
+  artist: {
+    name: "Abbas Kiarostami",
+    artworks_connection: {
+      pageInfo: {
+        startCursor: "YXJyYXljb25uZWN0aW9uOjA=",
+        endCursor: "YXJyYXljb25uZWN0aW9uOjc=",
+        hasNextPage: true,
+        hasPreviousPage: false,
+      },
+      edges: [
+        {
+          node: {
+            " $fragmentRefs": null,
+            id: "abbas-kiarostami-untitled-7",
+            _id: "5b2b745e9c18db204fc32e11",
+            title: "Untitled",
+          },
+        },
+        {
+          node: {
+            " $fragmentRefs": null,
+            id: "abbas-kiarostami-snow-no-12",
+            _id: "5a6b3d7d9c18db1962a94bc8",
+            title: "Snow No. 12",
+          },
+        },
+        {
+          node: {
+            " $fragmentRefs": null,
+            id: "abbas-kiarostami-snow-no-2",
+            _id: "5a6b3d70cd530e5037b87540",
+            title: "Snow No. 2",
+          },
+        },
+        {
+          node: {
+            " $fragmentRefs": null,
+            id: "abbas-kiarostami-snow-no-34",
+            _id: "5a6b3d7d9c18db1884f13c29",
+            title: "Snow No. 34",
+          },
+        },
+        {
+          node: {
+            " $fragmentRefs": null,
+            id: "abbas-kiarostami-road-series-30",
+            _id: "53be42857261692aaf510b00",
+            title: "Road Series 30",
+          },
+        },
+        {
+          node: {
+            " $fragmentRefs": null,
+            id: "abbas-kiarostami-road-series-2",
+            _id: "53be42857261696e662b0b00",
+            title: "Road Series 2",
+          },
+        },
+        {
+          node: {
+            " $fragmentRefs": null,
+            id: "abbas-kiarostami-snow-no-31",
+            _id: "5a6b3d7d8b3b810ecbcad2ba",
+            title: "Snow No. 31",
+          },
+        },
+        {
+          node: {
+            " $fragmentRefs": null,
+            id: "abbas-kiarostami-snow-no-3",
+            _id: "5ac44edbc9dc2452a47f6163",
+            title: "Snow No. 3",
+          },
+        },
+      ],
+    },
+  },
+}